+++ packages/ui/src/components/Slider/withAutoplayPlugin.ts
@@ -1,32 +1,45 @@
 import { KeenSliderPlugin } from 'keen-slider/react'
 
-const withAutoplay: KeenSliderPlugin = (slider) => {
-  let timeout
-  let mouseOver = false
-  function clearNextTimeout() {
-    clearTimeout(timeout)
-  }
-  function nextTimeout() {
-    clearTimeout(timeout)
-    if (mouseOver) return
-    timeout = setTimeout(() => {
-      slider.next()
-    }, 2000)
-  }
-  slider.on('created', () => {
-    slider.container.addEventListener('mouseover', () => {
-      mouseOver = true
-      clearNextTimeout()
-    })
-    slider.container.addEventListener('mouseout', () => {
-      mouseOver = false
+const withAutoplay =
+  (options: {
+    durationPerSlide: number
+    disablePauseOnHover?: boolean
+  }): KeenSliderPlugin =>
+  (slider) => {
+    const { disablePauseOnHover, durationPerSlide } = options
+
+    let timeout
+    let mouseOver = false
+
+    function clearNextTimeout() {
+      clearTimeout(timeout)
+    }
+    function nextTimeout() {
+      clearTimeout(timeout)
+      if (mouseOver && !disablePauseOnHover) return
+      timeout = setTimeout(() => {
+        slider.next()
+      }, durationPerSlide)
+    }
+
+    slider.on('created', () => {
+      if (!disablePauseOnHover) {
+        slider.container.addEventListener('mouseover', () => {
+          mouseOver = true
+          clearNextTimeout()
+        })
+        slider.container.addEventListener('mouseout', () => {
+          mouseOver = false
+          nextTimeout()
+        })
+      }
+
       nextTimeout()
     })
-    nextTimeout()
-  })
-  slider.on('dragStarted', clearNextTimeout)
-  slider.on('animationEnded', nextTimeout)
-  slider.on('updated', nextTimeout)
-}
+
+    slider.on('dragStarted', clearNextTimeout)
+    slider.on('animationEnded', nextTimeout)
+    slider.on('updated', nextTimeout)
+  }
 
 export default withAutoplay